Began to refactor Sails density into folders
Removed the `sails-density.cpp` and `sails-density.h` files to streamline the codebase. This also includes the deletion of all associated classes, methods, and functions related to density calculation, scoring, and MTZ file handling from these files.
